+++ eqcorrscan/core/match_filter/family.py
@@ -633,7 +633,7 @@ class Family(object):
             vertical_chans=vertical_chans, cores=cores,
             interpolate=interpolate, plot=plot, plotdir=plotdir,
             export_cc=export_cc, cc_dir=cc_dir,
-            check_full_seed=check_full_seed)
+            check_full_seed=check_full_seed, **kwargs)
         catalog_out = Catalog([ev for ev in picked_dict.values()])
         for detection_id, event in picked_dict.items():
             for pick in event.picks: