+++ eqcorrscan/utils/catalog_to_dd.py
@@ -354,7 +354,8 @@ def _compute_dt_correlations(catalog, master, min_link, event_id_mapper,
                         continue
                     correlation = ccc_out[i][j]
                     if interpolate:
-                        shift, cc_max = _xcorr_interp(correlation, dt=delta)
+                        shift, cc_max = _xcorr_interp(correlation, dt=delta,
+                                                      **kwargs)
                     else:
                         cc_max = np.amax(correlation)
                         shift = np.argmax(correlation) * delta
@@ -556,6 +557,8 @@ def compute_differential_times(catalog, correlation, stream_dict=None,
         min_cc=min_cc, stream_dict=stream_dict, extract_len=extract_len,
         pre_pick=pre_pick, shift_len=shift_len, interpolate=interpolate,
         max_workers=max_workers)
+    for key, value in kwargs.items():
+        correlation_kwargs.update({key: value})
     if correlation:
         for arg, name in correlation_kwargs.items():
             assert arg is not None, "{0} is required for correlation".format(