+++ CHANGES.md
@@ -1,6 +1,9 @@
 ## Current
 * core.match_filter.tribe
  - Detect now allows passing of pre-processed data
+* core.lag_calc._xcorr_interp
+ - CC-interpolation replaced with resampling (more robust), old method
+   deprecated. Use new method with use_new_resamp_method=True as **kwarg.
 * utils.correlate
  - Fast Matched Filter now supported natively for version >= 1.4.0
  - Only full correlation stacks are returned now (e.g. where fewer than than
@@ -21,6 +24,8 @@
    the old parallelization strategy across traces.
  - Now includes `all_horiz`-option that will correlate all matching horizontal
    channels no matter to which of these the S-pick is linking.
+* tribe, template, template_gen, archive_read, clustering: remove option to read
+  from seishub (deprecated in obspy).
 
 ## 0.4.3
 * core.match_filter